pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-broker"
version = "0.1.0"
description = "Compliance-aware LLM routing broker with OpenAI-compatible API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "litellm>=1.40.0",
    "routellm>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_broker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
